BUILD: apply make_base.py changes from 1.1.x also on develop

+++ sphinx/base/make_base.py
@@ -13,7 +13,8 @@ import sys
 from abc import ABCMeta, abstractmethod
 from glob import glob
 from tempfile import TemporaryDirectory
-from typing import Dict, Iterable, List, Set, Tuple, Type
+from typing import Any, Dict, Iterable, List, Optional, Set, Tuple, Type, TypeVar
+from weakref import ref
 
 from packaging import version as pkg_version
 
@@ -33,7 +34,7 @@ DIR_DOCS = os.path.join(DIR_REPO_ROOT, "docs")
 DIR_SPHINX_SOURCE = os.path.join(cwd, "source")
 DIR_SPHINX_AUX = os.path.join(cwd, "auxiliary")
 DIR_SPHINX_API_GENERATED = os.path.join(DIR_SPHINX_SOURCE, "apidoc")
-DIR_SPHINX_GSTART_GENERATED = os.path.join(DIR_SPHINX_SOURCE, "getting_started")
+DIR_SPHINX_GET_STARTED_GENERATED = os.path.join(DIR_SPHINX_SOURCE, "getting_started")
 DIR_SPHINX_BUILD = os.path.join(cwd, "build")
 DIR_SPHINX_BUILD_HTML = os.path.join(DIR_SPHINX_BUILD, "html")
 DIR_SPHINX_TEMPLATES = os.path.join(DIR_SPHINX_SOURCE, "_templates")
@@ -49,39 +50,65 @@ DIR_ALL_DOCS_VERSIONS = os.path.join(DIR_SPHINX_BUILD, "docs-version")
 # noinspection SpellCheckingInspection
 ENV_PYTHON_PATH = "PYTHONPATH"
 
+T = TypeVar("T")
 
-class Command(metaclass=ABCMeta):
+
+class CommandMeta(ABCMeta):
+    """
+    Meta-class for command classes.
+
+    Subsequent instantiations of a command class return the identical object.
+    """
+
+    def __init__(cls, *args: Any, **kwargs: Any) -> None:
+        super().__init__(*args, **kwargs)
+        cls.__instance_ref: Optional[ref] = None
+
+    def __call__(cls: Type[T], *args, **kwargs: Any) -> T:
+        """
+        Return the existing command instance, or create a new one if none exists yet.
+
+        :return: the command instance
+        """
+        if args or kwargs:
+            raise ValueError("command classes may not take any arguments")
+
+        cls: CommandMeta
+
+        if cls.__instance_ref:
+            obj = cls.__instance_ref()
+            if obj is not None:
+                return obj
+
+        instance = super(CommandMeta, cls).__call__()
+        cls.__instance_ref = ref(instance)
+        return instance
+
+
+class Command(metaclass=CommandMeta):
     """ Defines an available command that can be launched from this module."""
 
     __RE_CAMEL_TO_SNAKE = re.compile(r"(?<!^)(?=[A-Z])")
 
-    @classmethod
-    def get_name(cls) -> str:
-        try:
-            return cls.__name
-        except AttributeError:
-            cls.__name = cls.__RE_CAMEL_TO_SNAKE.sub("_", cls.__name__).lower()
-            return cls.__name
+    def __init__(self) -> None:
+        self.name = self.__RE_CAMEL_TO_SNAKE.sub("_", type(self).__name__).lower()
 
-    @classmethod
     @abstractmethod
-    def get_description(cls) -> str:
+    def get_description(self) -> str:
         pass
 
-    @classmethod
-    def get_dependencies(cls) -> Tuple[Type["Command"], ...]:
-        return ()
+    @abstractmethod
+    def get_dependencies(self) -> Tuple["Command", ...]:
+        pass
 
-    @classmethod
-    def get_prerequisites(cls) -> Iterable[Type["Command"]]:
-        dependencies_extended: List[Type["Command"]] = []
+    def get_prerequisites(self) -> Iterable["Command"]:
+        dependencies_extended: List[Command] = []
 
-        for dependency in cls.get_dependencies():
+        for dependency in self.get_dependencies():
             dependencies_inherited = dependency.get_dependencies()
-            if cls in dependencies_inherited:
+            if self in dependencies_inherited:
                 raise ValueError(
-                    f"circular dependency: {dependency.get_name()} "
-                    f"depends on {cls.get_name()}"
+                    f"circular dependency: {dependency.name} " f"depends on {self.name}"
                 )
             dependencies_extended.extend(
                 dependency
@@ -92,14 +119,12 @@ class Command(metaclass=ABCMeta):
 
         return dependencies_extended
 
-    @classmethod
-    def run(cls) -> None:
-        print(f"Running command {cls.get_name()} – {cls.get_description()}")
-        cls._run()
+    def run(self) -> None:
+        print(f"Running command {self.name} – {self.get_description()}")
+        self._run()
 
-    @classmethod
     @abstractmethod
-    def _run(cls) -> None:
+    def _run(self) -> None:
         pass
 
 
@@ -109,32 +134,30 @@ class Command(metaclass=ABCMeta):
 
 
 class Clean(Command):
-    @classmethod
-    def get_description(cls) -> str:
+    def get_description(self) -> str:
         return "remove Sphinx build output"
 
-    @classmethod
-    def _run(cls) -> None:
+    def get_dependencies(self) -> Tuple[Command, ...]:
+        return ()
+
+    def _run(self) -> None:
         if os.path.exists(DIR_SPHINX_BUILD):
             shutil.rmtree(path=DIR_SPHINX_BUILD)
         if os.path.exists(DIR_SPHINX_API_GENERATED):
             shutil.rmtree(path=DIR_SPHINX_API_GENERATED)
-        if os.path.exists(DIR_SPHINX_GSTART_GENERATED):
-            shutil.rmtree(path=DIR_SPHINX_GSTART_GENERATED)
+        if os.path.exists(DIR_SPHINX_GET_STARTED_GENERATED):
+            shutil.rmtree(path=DIR_SPHINX_GET_STARTED_GENERATED)
 
 
 class ApiDoc(Command):
-    @classmethod
-    def get_description(cls) -> str:
+    def get_description(self) -> str:
         return "generate Sphinx API documentation from sources"
 
-    @classmethod
-    def get_dependencies(cls) -> Tuple[Type["Command"], ...]:
+    def get_dependencies(self) -> Tuple[Command, ...]:
         # noinspection PyRedundantParentheses
-        return (Clean,)
+        return (Clean(),)
 
-    @classmethod
-    def _run(cls) -> None:
+    def _run(self) -> None:
         packages = [
             package for package in os.listdir(DIR_PACKAGE_SRC) if package[:1].isalnum()
         ]
@@ -173,20 +196,17 @@ class ApiDoc(Command):
 
 
 class GettingStartedDoc(Command):
-    @classmethod
-    def get_description(cls) -> str:
+    def get_description(self) -> str:
         return "generate getting started documentation from sources"
 
-    @classmethod
-    def get_dependencies(cls) -> Tuple[Type["Command"], ...]:
+    def get_dependencies(self) -> Tuple[Command, ...]:
         # noinspection PyRedundantParentheses
-        return (Clean,)
+        return (Clean(),)
 
-    @classmethod
-    def _run(cls) -> None:
+    def _run(self) -> None:
 
         # make dir if it does not exist
-        os.makedirs(DIR_SPHINX_GSTART_GENERATED, exist_ok=True)
+        os.makedirs(DIR_SPHINX_GET_STARTED_GENERATED, exist_ok=True)
 
         # open the rst readme file
         with open(os.path.join(DIR_REPO_ROOT, "README.rst"), "r") as file:
@@ -212,51 +232,27 @@ class GettingStartedDoc(Command):
             template_data = file.read()
 
         with open(
-            os.path.join(DIR_SPHINX_GSTART_GENERATED, "getting_started.rst"), "wt"
+            os.path.join(DIR_SPHINX_GET_STARTED_GENERATED, "getting_started.rst"),
+            "wt",
         ) as file:
             file.writelines(template_data)
             file.writelines(readme_data)
 
 
 class FetchPkgVersions(Command):
-    @classmethod
-    def get_description(cls) -> str:
+    def get_description(self) -> str:
         return "fetch available package versions with docs"
 
-    @classmethod
-    def get_dependencies(cls) -> Tuple[Type["Command"], ...]:
+    def get_dependencies(self) -> Tuple[Command, ...]:
         return ()
 
-    @classmethod
-    def _run(cls) -> None:
-        os.makedirs(DIR_SPHINX_BUILD, exist_ok=True)
-        start_from_version_tag = "1.0.1"
-        sp = subprocess.run(
-            args='git tag -l "*.*.*"', shell=True, check=True, stdout=subprocess.PIPE
-        )
-        version_tags = sp.stdout.decode("UTF-8").split("\n")
-        version_tags = [
-            vt for vt in version_tags if vt != "" and vt >= start_from_version_tag
-        ]
-
-        # add version currently build into version_tags
-        version_built = get_package_version()
-
-        if version_built not in version_tags:
-            version_tags.append(version_built)
-
-        version_tags.sort()
-        version_tags.reverse()
-        version_tags_non_rc = [vt for vt in version_tags if "rc" not in vt]
-        latest_non_rc_version = version_tags_non_rc[0]
-
-        print("Found the following version tags: ", version_tags)
-        print("Latest non-RC version: ", latest_non_rc_version)
+    def _run(self) -> None:
+        versions = Versions()
 
         version_data = {
-            "current": latest_non_rc_version,
-            "non_rc": version_tags_non_rc,
-            "all": version_tags,
+            "current": str(versions.latest_stable_version),
+            "non_rc": list(map(str, versions.version_tags_stable)),
+            "all": list(map(str, versions.version_tags)),
         }
 
         version_data_as_js = (
@@ -270,34 +266,43 @@ class FetchPkgVersions(Command):
 
 
 class PrepareDocsDeployment(Command):
-    @classmethod
-    def get_description(cls) -> str:
+    def get_description(self) -> str:
         return "update versions of rendered documentation"
 
-    @classmethod
-    def get_dependencies(cls) -> Tuple[Type["Command"], ...]:
+    def get_dependencies(self) -> Tuple[Command, ...]:
         return ()
 
-    @classmethod
-    def _run(cls) -> None:
-        assert is_azure_build(), "Only implemented for Azure Pipelines"
+    def _run(self) -> None:
+        if not is_azure_build():
+            raise RuntimeError("only implemented for Azure Pipelines")
+
+        # get the current version of the package
+        current_version: pkg_version.Version = get_package_version()
+
+        # copy new the docs version to the deployment path
+        if current_version == Versions().latest_stable_version:
+            # only copy to deployment path if our version is the latest stable release
+
+            # remove docs build currently deployed, except for the docs versions folder
+            if os.path.exists(os.path.join(DIR_DOCS, "docs-version")):
+                with TemporaryDirectory() as DIR_TMP:
+                    shutil.move(src=os.path.join(DIR_DOCS, "docs-version"), dst=DIR_TMP)
+                    shutil.rmtree(path=DIR_DOCS)
+                    os.makedirs(DIR_DOCS)
+                    shutil.move(src=os.path.join(DIR_TMP, "docs-version"), dst=DIR_DOCS)
+            else:
+                os.makedirs(DIR_DOCS, exist_ok=True)
+
+            # copy new docs version to deployment path
+            shutil.copytree(src=DIR_SPHINX_BUILD_HTML, dst=DIR_DOCS, dirs_exist_ok=True)
+
         # get current version of package in the form of folder/URL name (e.g., "1-0-0")
-        current_version = version_string_to_url(get_package_version())
-        # remove docs build currently deployed, except for the docs versions folder
-        if os.path.exists(os.path.join(DIR_DOCS, "docs-version")):
-            with TemporaryDirectory() as DIR_TMP:
-                shutil.move(src=os.path.join(DIR_DOCS, "docs-version"), dst=DIR_TMP)
-                shutil.rmtree(path=DIR_DOCS)
-                os.makedirs(DIR_DOCS)
-                shutil.move(src=os.path.join(DIR_TMP, "docs-version"), dst=DIR_DOCS)
-
-        # copy new docs version to deployment path
-        os.makedirs(DIR_DOCS, exist_ok=True)
-        shutil.copytree(src=DIR_SPHINX_BUILD_HTML, dst=DIR_DOCS, dirs_exist_ok=True)
-
+        current_version_path = os.path.join(
+            DIR_DOCS, "docs-version", version_string_to_url(current_version)
+        )
         # update latest version in docs history
-        if os.path.exists(os.path.join(DIR_DOCS, "docs-version", current_version)):
-            shutil.rmtree(path=os.path.join(DIR_DOCS, "docs-version", current_version))
+        if os.path.exists(current_version_path):
+            shutil.rmtree(path=current_version_path)
         shutil.copytree(
             src=DIR_ALL_DOCS_VERSIONS,
             dst=os.path.join(DIR_DOCS, "docs-version"),
@@ -316,22 +321,20 @@ class PrepareDocsDeployment(Command):
             os.remove(os.path.join(DIR_DOCS, ".buildinfo"))
 
         # create empty file to signal that no GitHub auto-rendering is required
+        # noinspection SpellCheckingInspection
         open(os.path.join(DIR_DOCS, ".nojekyll"), "a").close()
 
         print("Docs moved to ./docs and historic versions updated")
 
 
 class Html(Command):
-    @classmethod
-    def get_description(cls) -> str:
+    def get_description(self) -> str:
         return "build Sphinx docs as HTML"
 
-    @classmethod
-    def get_dependencies(cls) -> Tuple[Type["Command"], ...]:
-        return Clean, FetchPkgVersions, ApiDoc, GettingStartedDoc
+    def get_dependencies(self) -> Tuple[Command, ...]:
+        return Clean(), FetchPkgVersions(), ApiDoc(), GettingStartedDoc()
 
-    @classmethod
-    def _run(cls) -> None:
+    def _run(self) -> None:
 
         check_sphinx_version()
 
@@ -353,7 +356,7 @@ class Html(Command):
         sys.path.append(DIR_MAKE_BASE)
 
         # create copy of this build for the docs archive
-        version_built = get_package_version()
+        version_built: pkg_version.Version = get_package_version()
         dir_path_this_build = os.path.join(
             DIR_ALL_DOCS_VERSIONS, version_string_to_url(version_built)
         )
@@ -376,13 +379,58 @@ class Html(Command):
 
 
 class Help(Command):
-    @classmethod
-    def get_description(cls) -> str:
+    def get_description(self) -> str:
         return "print this help message"
 
-    @classmethod
-    def _run(cls) -> None:
+    def get_dependencies(self) -> Tuple[Command, ...]:
+        return ()
+
+    def _run(self) -> None:
         print_usage()
+
+
+class Versions:
+    """
+    Helper class that lists all versions that have already been released.
+    """
+
+    INITIAL_VERSION_TAG = pkg_version.parse("1.0.1")
+
+    def __init__(self) -> None:
+        os.makedirs(DIR_SPHINX_BUILD, exist_ok=True)
+        start_from_version_tag: pkg_version.Version = Versions.INITIAL_VERSION_TAG
+        sp = subprocess.run(
+            args='git tag -l "*.*.*"', shell=True, check=True, stdout=subprocess.PIPE
+        )
+        version_tags: List[pkg_version.Version] = [
+            version_tag
+            for version_tag in (
+                pkg_version.parse(version_string)
+                for version_string in sp.stdout.decode("UTF-8").split("\n")
+                if version_string
+            )
+            if version_tag >= start_from_version_tag
+        ]
+
+        # add version currently build into version_tags
+        version_built: pkg_version.Version = get_package_version()
+
+        if version_built not in version_tags:
+            version_tags.append(version_built)
+
+        version_tags.sort()
+        version_tags.reverse()
+        version_tags_stable: List[pkg_version.Version] = [
+            vt for vt in version_tags if not (vt.is_prerelease or vt.is_devrelease)
+        ]
+        latest_stable_version: pkg_version.Version = version_tags_stable[0]
+
+        print(f"Found versions: {', '.join(map(str, version_tags))}")
+        print("Latest stable version: ", latest_stable_version)
+
+        self.version_tags = version_tags
+        self.version_tags_stable = version_tags_stable
+        self.latest_stable_version = latest_stable_version
 
 
 def make(*, modules: List[str]) -> None:
@@ -411,11 +459,11 @@ def make(*, modules: List[str]) -> None:
     os.environ[ENV_PYTHON_PATH] = os.pathsep.join(module_paths)
 
     # run all given commands:
-    executed_commands: Set[Type[Command]] = set()
+    executed_commands: Set[Command] = set()
 
     for next_command_name in commands_passed:
 
-        next_command: Type[Command] = available_commands[next_command_name]
+        next_command: Command = available_commands[next_command_name]
 
         for prerequisite_command in next_command.get_prerequisites():
 
@@ -437,7 +485,7 @@ def quote_path(path: str) -> str:
         return path
 
 
-def get_package_version() -> str:
+def get_package_version() -> pkg_version.Version:
     """
     Retrieve the package version for the project from __init__ or _version
     """
@@ -459,7 +507,7 @@ def get_package_version() -> str:
     # noinspection PyUnresolvedReferences
     spec.loader.exec_module(version_module)
     # noinspection PyUnresolvedReferences
-    return version_module.__version__
+    return pkg_version.parse(version_module.__version__)
 
 
 def is_azure_build() -> bool:
@@ -469,13 +517,13 @@ def is_azure_build() -> bool:
     return "BUILD_REASON" in os.environ
 
 
-def version_string_to_url(version: str) -> str:
+def version_string_to_url(version: pkg_version.Version) -> str:
     """
     Make a Python package version string safe for URLs/folders.
 
     Our convention is to only replace all dots with dashes.
     """
-    return version.replace(".", "-")
+    return str(version).replace(".", "-")
 
 
 def check_sphinx_version() -> None:
@@ -499,15 +547,15 @@ Available program arguments:
     print(usage)
 
 
-available_commands: Dict[str, Type[Command]] = {
-    cmd.get_name(): cmd
+available_commands: Dict[str, Command] = {
+    cmd.name: cmd
     for cmd in (
-        Clean,
-        ApiDoc,
-        GettingStartedDoc,
-        Html,
-        Help,
-        FetchPkgVersions,
-        PrepareDocsDeployment,
+        Clean(),
+        ApiDoc(),
+        GettingStartedDoc(),
+        Html(),
+        Help(),
+        FetchPkgVersions(),
+        PrepareDocsDeployment(),
     )
 }
